Use table instead of dl for RDoc label/note lists
dl is very difficult to style correctly to get the same visual
result as table.  You can get close, but you have to make
sacrifices such as using a fixed width for the label.  It's much
simpler and looks much better to just use a table, which is how
GitHub styles the RDoc.